Move named_params.rs into statement.rs, greatly reducing the StatementCrateImpl trait size.
